RDKB-57868: Return appropriate error/success return value for switch channel.
Reason for change: Due to incorrect return value OneWifi was not retrying
channel change when the failure happened. This commit is to ensure that
wifi-hal returns appropriate value. In case of failure retval, OneWifi
can retry the channel change.

Test Procedure: Ensured that the config change is reflected in radio
subdoc as part of successful channel change request to agent.

Risks: Low
Priority: P1
Signed-off-by: Amarnath Hullur Subramanyam <amarnath_hullursubramanyam@comcast.com>
Change-Id: I57e4fd54591c82673dbf6ce2212ef929b4e7f04b